% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creds.R
\name{add_creds}
\alias{add_creds}
\title{Add user credentials to the keyring.}
\usage{
add_creds(.report = NULL)
}
\arguments{
\item{.report}{A list -- report returned from hRvstAPI::check_creds()}
}
\description{
This is intended to be a helper function for setting credentials.
}
