% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{create_db}
\alias{create_db}
\title{Create a SQLite database to hold Harvest API v2 request data.}
\usage{
create_db(db_path = NULL, rds_path = NULL)
}
\arguments{
\item{db_path}{A string -- file path where a local .sqlite file containing Harvest API data should be created.}

\item{rds_path}{A string -- file path where a local .rds file containing Harvest API data exists.}
}
\description{
Create a SQLite database to hold Harvest API v2 request data.
}
