% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{create_rds}
\alias{create_rds}
\title{Download Harvest API v2 requests into a local file.}
\usage{
create_rds(..., rds_path = NULL, sql_colnames = TRUE)
}
\arguments{
\item{rds_path}{A string -- file path where .rds file containing Harvest API data should be created.}

\item{sql_colnames}{}
}
\description{
Download Harvest API v2 requests into a local file.
}
