% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{get_keys}
\alias{get_keys}
\title{Get the unique set of key values that already exist in a table.}
\usage{
get_keys(db_connection, tbl, key = NULL)
}
\arguments{
\item{key}{}
}
\description{
Get the unique set of key values that already exist in a table.
}
