% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req.R
\name{harvest_req}
\alias{harvest_req}
\title{Perform a request for a Harvest API resource.}
\usage{
harvest_req(
  resource = NULL,
  all_pages = TRUE,
  base_url = NULL,
  headers = NULL,
  is_active = NULL,
  ...
)
}
\arguments{
\item{resource}{A string -- modifies the URL path to point to a specific Harvest API v2 resource. Currently available options:
\itemize{
\item clients
\item projects
\item tasks
\item users
\item user assignments
\item task assignments
\item time entries
\item budget reports
\item time reports
}}

\item{all_pages}{A boolean -- should all pages be gathered for a requested resource? (default value TRUE)}

\item{base_url}{-- A string -- the common URL component for Harvest API v2 requests, passed to \code{\link[=harvest_GET]{harvest_GET()}}.}

\item{headers}{A list -- the headers required for authentication of each Harvest API v2 request, passed to \code{\link[=harvest_GET]{harvest_GET()}}.}

\item{...}{A named list of (optional) additional query parameters, passed to \code{\link[=harvest_GET]{harvest_GET()}}.}
}
\value{
If request is successful (i.e. the request was successfully
performed and a response with HTTP status code <400 was recieved), an HTTP
\code{\link[httr2:response]{httr2::response()}}; otherwise throws an error.
}
\description{
This is the primary function used to gather resources from the
Harvest API v2. It uses \code{\link[=harvest_GET]{harvest_GET()}} to construct
an \code{\link[httr2:request]{httr2::request()}} and then performs that request after
modifying the URL.
}
\details{
If only a specific page is required, the \verb{page = [n]} parameter
can be passed to \code{...} to be added to the query parameters.
Adding this query parameter will automatically set \emph{all_pages}
to FALSE.
}
\seealso{
\href{https://help.getharvest.com/api-v2/introduction/overview/general/#rate-limiting}{Harvest API V2 Documentation | Rate Limiting}
}
