% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creds.R
\name{harvest_token}
\alias{harvest_token}
\title{Return the user's Harvest API v2 personal access token.}
\usage{
harvest_token()
}
\value{
A string -- the user personal access token.
}
\description{
Return the user's Harvest API v2 personal access token.
}
\seealso{
\href{https://help.getharvest.com/api-v2/authentication-api/authentication/authentication/}{Harvest API V2 Documentation | Authentication}
}
