% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{hrvst_db}
\alias{hrvst_db}
\title{Transfer local Harvest data into SQLite database.}
\usage{
hrvst_db(rds_file = NULL, db_file = NULL)
}
\arguments{
\item{rds_file}{A string -- file path where .rds file containing Harvest API data is located.}

\item{db_file}{A string -- file path where .sqlite file containing Harvest API data should be/is located.}
}
\description{
Transfer local Harvest data into SQLite database.
}
