% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req.R
\name{hrvst_GET}
\alias{hrvst_GET}
\title{Create an authenticated GET request with headers and queries.}
\usage{
hrvst_GET(
  url = NULL,
  headers = NULL,
  is_active = NULL,
  updated_since = NULL,
  from = NULL,
  to = NULL,
  ...
)
}
\arguments{
\item{url}{A string -- the URL component for Harvest API v2 requests (NULL will default to \link{.url}).}

\item{headers}{A list -- the headers required for authentication of each Harvest API v2 request: 'Harvest-Account-Id', 'Authorization' (PAT), and 'User-Agent' (this package GitHub repo, \link{.agent}).}

\item{is_active}{A string -- 'true' or 'false' (boolean per Harvest API requirement) should only active ('true') or inactive ('false') records be requested; default NULL downloads all records.}

\item{updated_since}{A string -- a date or date/time value (in ISO8601 character format) indicated how far back in time to retrieve records. Not supported by all of the Harvest v2 APIs.}

\item{from}{TODO}

\item{to}{TODO}

\item{...}{Key-value pairs -- optional additional query parameters.}
}
\value{
An HTTP response: an S3 list with class \code{httr2_request}.
}
\description{
The basic headers that are automatically set in this function are
\code{Harvest-Account-Id}, \code{Authorization}, and \code{User-Agent}. Other headers
can be added by the user and will be concatenated, but currently
these important headers are set by the \code{hRvstAPI::hrvst_acct_id()},
\code{hRvstAPI::hrvst_token()}, and \code{hRvstAPI::.agent} objects.
}
\seealso{
\code{\link[httr2:request]{httr2::request()}}
}
