% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{hrvst_rds}
\alias{hrvst_rds}
\title{Download Harvest API v2 requests into a local file.}
\usage{
hrvst_rds(..., rds_file = NULL, sql_colnames = TRUE)
}
\arguments{
\item{rds_file}{A string -- file path where .rds file containing Harvest API data should be created.}

\item{sql_colnames}{}
}
\description{
Download Harvest API v2 requests into a local file.
}
