% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req.R
\name{hrvst_req}
\alias{hrvst_req}
\title{Perform a request for a Harvest API resource.}
\usage{
hrvst_req(
  resource = NULL,
  all_pages = TRUE,
  base_url = NULL,
  headers = NULL,
  is_active = "true",
  updated_since = NULL,
  weeks_ago = NULL,
  from = NULL,
  to = NULL,
  ...
)
}
\arguments{
\item{resource}{A string -- modifies the URL path to point to a specific Harvest API v2 resource. Currently available options:
\itemize{
\item \strong{users}
(\href{https://help.getharvest.com/api-v2/users-api/}{Users API})
\item \strong{clients}
(\href{https://help.getharvest.com/api-v2/clients-api/}{Clients API})
\item \strong{projects}
(\href{https://help.getharvest.com/api-v2/projects-api/}{Projects API})
\item \strong{tasks}
(\href{https://help.getharvest.com/api-v2/tasks-api/}{Tasks API})
\item \strong{project assignments}
(via \href{https://help.getharvest.com/api-v2/users-api/}{Users API})
\emph{i.e Which projects belong to which clients.}
\item \strong{user assignments}
(via \href{https://help.getharvest.com/api-v2/projects-api/}{Projects API})
\emph{i.e. How users are assigned to each projects.}
\item \strong{task assignments}
(via \href{https://help.getharvest.com/api-v2/projects-api/}{Projects API})
\emph{i.e. What tasks are assigned to each project.}
\item \strong{time entries}
(via \href{https://help.getharvest.com/api-v2/timesheets-api/}{Timesheets API})
\item \strong{budget reports}
(via \href{https://help.getharvest.com/api-v2/reports-api/}{Reports API})
\item \strong{time reports}
(via \href{https://help.getharvest.com/api-v2/reports-api/}{Reports API})
}}

\item{all_pages}{A boolean -- should all pages be gathered for a requested resource? (default value TRUE)}

\item{base_url}{-- A string -- the common URL component for Harvest API v2 requests, passed to \code{\link[=hrvst_GET]{hrvst_GET()}}.}

\item{headers}{A list -- passed to \code{\link[=hrvst_GET]{hrvst_GET()}}.}

\item{is_active}{A string -- 'true' or 'false' passed to \code{\link[=hrvst_GET]{hrvst_GET()}}.}

\item{updated_since}{A string -- passed to \code{\link[=hrvst_GET]{hrvst_GET()}}.}

\item{weeks_ago}{An integer -- overrides \emph{updated_since} argument; the number of weeks prior to the current week to convert to a date and pass to the 'updated_since' query parameter.}

\item{from}{TODO}

\item{to}{TODO}

\item{...}{Key-value pairs -- passed to \code{\link[=hrvst_GET]{hrvst_GET()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} of all response content.
}
\description{
This is the primary function used to gather resources from the
Harvest API v2. It uses \code{\link[=hrvst_GET]{hrvst_GET()}} to construct
an \code{\link[httr2:request]{httr2::request()}} and then performs that request after
modifying the URL.

\emph{Requests are automatically throttled to comply with
Harvest API v2 constraints.}
}
\details{
If only a specific page is required, the \verb{page = [n]} parameter
can be passed to \code{...} to be added to the query parameters.
Adding this query parameter will automatically set \emph{all_pages}
to FALSE.
}
\seealso{
\href{https://help.getharvest.com/api-v2/introduction/overview/general/#rate-limiting}{Harvest API V2 Documentation | Rate Limiting}
}
