% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{key_col}
\alias{key_col}
\title{Get the name of the primary key column for a table.}
\usage{
key_col(db_connection, tbl)
}
\arguments{
\item{tbl}{}
}
\description{
Get the name of the primary key column for a table.
}
