% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\docType{data}
\name{params}
\alias{params}
\alias{.service}
\alias{.url}
\alias{.agent}
\alias{.rds_path}
\alias{.db_path}
\title{hRvstAPI parameters}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
.service

.url

.agent

.rds_path

.db_path
}
\description{
These are common values used by hRvstAPI functions like the base URL for
the Harvest API v2, and the name of the hRvstAPI GitHub repo for
a user agent to be logged with each GET request made to the Harvest API.
In the future these may be set as options to give the user even more control
over this wrapper.
}
\keyword{datasets}
