% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{query_db}
\alias{query_db}
\title{Query a database and collect/return the result.}
\usage{
query_db(db_connection, query_string, ...)
}
\arguments{
\item{...}{}
}
\description{
Query a database and collect/return the result.
}
