% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creds.R
\name{report_creds}
\alias{report_creds}
\title{See if and where user credentials exist.}
\usage{
report_creds()
}
\value{
A list -- variables related to credential existence.
}
\description{
Useful for informing subsequent user action within an application.
}
