% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creds.R
\name{retrieve_creds}
\alias{retrieve_creds}
\title{Retrieve user credentials from the keyring to environment.}
\usage{
retrieve_creds()
}
\description{
This is intended to be a helper function for setting credentials.
}
