% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{update_db}
\alias{update_db}
\title{Update the local SQLite database with new Harvest API v2 request data.}
\usage{
update_db(db_path = NULL, rds_path = NULL)
}
\arguments{
\item{db_path}{A string -- file path where a local .sqlite file containing Harvest API data should be updated.}

\item{rds_path}{A string -- file path where a local .rds file containing Harvest API data should exist.}
}
\description{
Update the local SQLite database with new Harvest API v2 request data.
}
