% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req.R
\name{hrvst_GET}
\alias{hrvst_GET}
\title{Create an authenticated GET request with headers and queries.}
\usage{
hrvst_GET(
  url = NULL,
  headers = NULL,
  is_active = NULL,
  updated_since = NULL,
  from = NULL,
  to = NULL,
  ...
)
}
\arguments{
\item{url}{A string -- the URL component for Harvest API v2 requests. (NULL will default to hRvstAPI::.url)}

\item{headers}{A list -- the headers required for authentication of each Harvest API v2 request.}

\item{is_active}{A string -- 'true' or 'false', boolean per Harvest API requirement.}

\item{to}{}

\item{...}{A named list of (optional) additional query parameters.}
}
\value{
An HTTP response: an S3 list with class \code{httr2_request}.
}
\description{
The basic headers that are automatically set in this function are
\code{Harvest-Account-Id}, \code{Authorization}, and \code{User-Agent}. Other headers
can be added by the user and will be concatenated, but currently
these important headers are set by the \code{hRvstAPI::hrvst_acct_id()},
\code{hRvstAPI::hrvst_token()}, and \code{hRvstAPI::.agent} objects.
}
\seealso{
\code{\link[httr2:request]{httr2::request()}}
}
