% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{query_db}
\alias{query_db}
\title{Query a database and collect/return the result.}
\usage{
query_db(db_connection, query_string, ...)
}
\arguments{
\item{db_connection}{A DBI compatible connection object -- the database containing Harvest API v2 request data.}

\item{query_string}{A string -- the query to be executed on \emph{db_connection}.}

\item{...}{}
}
\description{
Query a database and collect/return the result.
}
