% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{query_db}
\alias{query_db}
\title{Query the local database and collect/return the result.}
\usage{
query_db(query_string, ..., .db_loc = hRvstAPI::.db_path)
}
\arguments{
\item{query_string}{A string -- the query to be executed.}

\item{...}{Key-value pairs -- additional arguments for interpolation passed to \code{\link[glue:glue_sql]{glue::glue_sql()}}.}

\item{.db_loc}{A string -- file path where a local .sqlite file containing Harvest API data should exist.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} of data collected from the query.
}
\description{
Query the local database and collect/return the result.
}
